stance-detection/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
numpy==1.26.3
openai==1.13.3
pandas==2.2.0
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
pywatchman==1.4.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.2.0
